src/epics/index.js -> src/redux/epics/index.js
